refactor(deps): extract ECMA-262 abstract operations into packages/ecma262-abstract
Split 262.ts into individual files under a new top-level composite package:
- ToString, ToNumber, TimeClip, ToObject, SameValue, ArrayCreate
- HasOwnProperty, Type, DateOperations, OrdinaryHasInstance, ToPrimitive

Each exported function gets its own file. Private helpers (ToInteger,
IsCallable, OrdinaryToPrimitive, mod) stay with their consumers.

ecma262-abstract has zero dependencies on ecma402-abstract (no circular dep).
ZERO constant and invariant() are inlined since they're trivial.

Also auto-includes root package.json in all vitest targets for #packages/*
subpath import resolution.
