+++ packages/ecma262-abstract/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "paths": {
+      "#packages/*": ["../../packages/*"]
+    }
+  },
+  "extends": "../../tsconfig.json"
+}
