+++ packages/ecma402-abstract/BUILD.bazel
@@ -14,6 +14,7 @@ SRC_DEPS = [
     "//:node_modules/@formatjs/bigdecimal",
     "//:node_modules/@formatjs/fast-memoize",
     "//:node_modules/@formatjs/intl-localematcher",
+    "//packages/ecma262-abstract",
 ]
 
 ts_compile(
