+++ packages/ecma402-abstract/CoerceOptionsToObject.ts
@@ -1,4 +1,4 @@
-import {ToObject} from './262.js'
+import {ToObject} from '#packages/ecma262-abstract/ToObject.js'
 
 /**
  * https://tc39.es/ecma402/#sec-coerceoptionstoobject
