+++ packages/ecma402-abstract/DateTimeFormat/FormatDateTimePattern.ts
@@ -1,4 +1,4 @@
-import {TimeClip} from '#packages/ecma402-abstract/262.js'
+import {TimeClip} from '#packages/ecma262-abstract/TimeClip.js'
 import {
   type DateTimeFormat,
   type DateTimeFormatLocaleInternalData,
