+++ packages/ecma402-abstract/DateTimeFormat/FormatDateTimeToParts.ts
@@ -1,4 +1,4 @@
-import {ArrayCreate} from '#packages/ecma402-abstract/262.js'
+import {ArrayCreate} from '#packages/ecma262-abstract/ArrayCreate.js'
 import {type IntlDateTimeFormatPart} from '#packages/ecma402-abstract/types/date-time.js'
 import type Decimal from '@formatjs/bigdecimal'
 import {PartitionDateTimePattern} from './PartitionDateTimePattern.js'
