+++ packages/ecma402-abstract/DateTimeFormat/PartitionDateTimeRangePattern.ts
@@ -1,4 +1,5 @@
-import {SameValue, TimeClip} from '#packages/ecma402-abstract/262.js'
+import {SameValue} from '#packages/ecma262-abstract/SameValue.js'
+import {TimeClip} from '#packages/ecma262-abstract/TimeClip.js'
 import {PartitionPattern} from '#packages/ecma402-abstract/PartitionPattern.js'
 import {
   type IntlDateTimeFormatPart,
