+++ packages/ecma402-abstract/DurationFormat/ToIntegerIfIntegral.ts
@@ -1,4 +1,4 @@
-import {ToNumber} from '#packages/ecma402-abstract/262.js'
+import {ToNumber} from '#packages/ecma262-abstract/ToNumber.js'
 import {invariant} from '#packages/ecma402-abstract/utils.js'
 
 export function ToIntegerIfIntegral(arg: any): number {
