+++ packages/ecma402-abstract/PluralRules/tsconfig.json
@@ -8,6 +8,9 @@
   "extends": "../../../tsconfig.json",
   "references": [
     {
+      "path": "../../ecma262-abstract/tsconfig.json"
+    },
+    {
       "path": "../tsconfig.json"
     },
     {
