+++ packages/ecma402-abstract/RelativeTimeFormat/BUILD.bazel
@@ -5,6 +5,7 @@ PACKAGE_NAME = "RelativeTimeFormat"
 SRCS = glob(["*.ts"])
 
 SRC_DEPS = [
+    "//packages/ecma262-abstract",
     "//packages/ecma402-abstract",
     "//packages/ecma402-abstract/types",
 ]
@@ -20,6 +21,7 @@ ts_compile(
 generate_ide_tsconfig_json(
     composite = True,
     project_references = [
+        "//packages/ecma262-abstract",
         "//packages/ecma402-abstract",
         "//packages/ecma402-abstract/types",
     ],
