+++ packages/intl-datetimeformat/BUILD.bazel
@@ -48,6 +48,7 @@ TESTS = glob([
 ])
 
 SRC_DEPS = [
+    "//packages/ecma262-abstract",
     "//packages/ecma402-abstract",
     "//packages/ecma402-abstract/DateTimeFormat",
     "//packages/ecma402-abstract/types",
@@ -767,6 +768,7 @@ ts_run_binary(
         "//:node_modules/lodash-es",
         "//:node_modules/minimist",
         "//:root_package_json",
+        "//packages/ecma262-abstract",
         "//packages/ecma402-abstract",
         "//packages/ecma402-abstract/DateTimeFormat",
         "//packages/ecma402-abstract/types",
@@ -875,6 +877,7 @@ generate_src_file(
         ":zdumps",
         "//:node_modules/json-stable-stringify",
         "//:root_package_json",
+        "//packages/ecma262-abstract",
         "//packages/ecma402-abstract",
         "//packages/ecma402-abstract/DateTimeFormat",
         "//packages/ecma402-abstract/types",
@@ -895,6 +898,7 @@ ts_run_binary(
         "//:node_modules/json-stable-stringify",
         "//:node_modules/minimist",
         "//:root_package_json",
+        "//packages/ecma262-abstract",
         "//packages/ecma402-abstract",
         "//packages/ecma402-abstract/DateTimeFormat",
         "//packages/ecma402-abstract/types",
@@ -924,6 +928,7 @@ ts_run_binary(
         "//:node_modules/json-stable-stringify",
         "//:node_modules/minimist",
         "//:root_package_json",
+        "//packages/ecma262-abstract",
         "//packages/ecma402-abstract",
         "//packages/ecma402-abstract/DateTimeFormat",
         "//packages/ecma402-abstract/types",
