+++ packages/intl-displaynames/BUILD.bazel
@@ -34,6 +34,7 @@ TESTS = glob([
 ])
 
 SRC_DEPS = [
+    "//packages/ecma262-abstract",
     "//packages/ecma402-abstract",
     "//packages/ecma402-abstract/types",
     "//packages/ecma402-abstract/DisplayNames",
@@ -726,6 +727,7 @@ ts_run_binary(
         "//:node_modules/fs-extra",
         "//:node_modules/json-stable-stringify",
         "//:node_modules/minimist",
+        "//packages/ecma262-abstract",
         "//packages/ecma402-abstract",
         "//packages/ecma402-abstract/types",
     ],
